refactor: split cell into its own component
