+++ src/components/objectListing/objectListing.component.tsx
@@ -16,8 +16,8 @@ import { useSkylarkSearchableObjectTypes } from "src/hooks/useSkylarkObjectTypes
 
 import { CreateButtons } from "./createButtons";
 import { RowActions } from "./rowActions";
-import { Search } from "./search/search.component";
-import { Table } from "./table";
+import { Search } from "./search";
+import { Table, TableCell } from "./table";
 
 const orderedKeys = ["__typename", "title", "name", "uid", "external_id"];
 
@@ -41,38 +41,7 @@ const createColumns = (
   const createdColumns = columns.map((column) =>
     columnHelper.accessor(column, {
       header: formatColumnHeader(column),
-      cell: ({ getValue, row, column, table }) => {
-        const initialValue = getValue();
-        // We need to keep and update the state of the cell normally
-        // eslint-disable-next-line react-hooks/rules-of-hooks
-        const [value, setValue] = useState(initialValue);
-
-        // When the input is blurred, we'll call our table meta's updateData function
-        // const onBlur = () => {
-        //   table.options.meta?.updateData(index, id, value);
-        // };
-
-        // If the initialValue is changed external, sync it up with our state
-        // eslint-disable-next-line react-hooks/rules-of-hooks
-        useEffect(() => {
-          setValue(initialValue);
-        }, [initialValue]);
-
-        return row.id === table.options.meta?.rowInEditMode ? (
-          <input
-            value={value as string}
-            onChange={(e) => setValue(e.target.value)}
-            className={clsx(
-              "w-full border-b-2 border-brand-primary py-1 outline-none disabled:border-none disabled:border-manatee-200 disabled:text-manatee-500",
-              initialValue !== value && "border-warning",
-              value === "" && initialValue !== "" && "border-error",
-            )}
-            disabled={column.id === "uid"}
-          />
-        ) : (
-          (initialValue as string)
-        );
-      },
+      cell: (props) => <TableCell {...props} />,
     }),
   );
 
@@ -153,7 +122,7 @@ export const ObjectList = ({
 
   const parsedColumns = useMemo(
     () => createColumns(sortedHeaders, { withObjectSelect, withObjectEdit }),
-    [sortedHeaders, withObjectSelect],
+    [sortedHeaders, withObjectEdit, withObjectSelect],
   );
 
   const [rowInEditMode, setRowInEditMode] = useState("");
