+++ src/components/objectListing/search/index.tsx
@@ -1,0 +1,1 @@
+export * from "./search.component";
