+++ src/components/objectListing/table/cell/index.tsx
@@ -1,0 +1,1 @@
+export * from "./cell.component";
