+++ src/components/objectListing/table/index.tsx
@@ -1,1 +1,2 @@
 export * from "./table.component";
+export * from "./cell";
